Adding a flag 'runtype'  to set the oses for PR build
The flag `runtype` can be set either to build or test
If its set to test the dockerimages are only built for
ubuntu, alpine and ubi if the script is running on a
x86_64 Linux machine to build OpenJ9 images.

The github workflows are using `build_latest.sh` script
to generate the docker images on each PR. Recent changes
to OpenJ9 docker images added the SCC generation feature
to it which is taking up quite a bit of time to build the
docker images.

Signed-off-by: bharathappali <bharath.appali@gmail.com>